gadgets/run: Move common functions to helpers.go
